Moved ElementType components to ElementType folder
